requirements.txt
# Core
python-dotenv==1.1.0
click==8.2.0

# Data
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.6.1
joblib==1.5.0
matplotlib==3.10.3
seaborn==0.13.2

# ML
xgboost==1.7.6
optuna==4.3.0

# Backend
fastapi==0.111.0
uvicorn==0.29.0

# Frontend
streamlit==1.45.1

# Feature Store
hopsworks==4.2.4
confluent-kafka==2.3.0  # El <3 ya no hace falta

